requirements.txt
requests==2.28.2
urllib3==1.26.15
python-dotenv==1.0.0
python-telegram-bot==13.15
